wip: move nft cli tests to e2e tests
