+++ tests/e2e/nft/grpc.go
@@ -1,4 +1,4 @@
-package testutil
+package nft
 
 import (
 	"fmt"
