+++ tests/e2e/nft/test_helper.go
@@ -1,4 +1,4 @@
-package testutil
+package nft
 
 import (
 	"fmt"
