+++ tests/e2e/nft/tx.go
@@ -1,4 +1,4 @@
-package testutil
+package nft
 
 import (
 	"fmt"
